Call startInput on return from sleep mode
One manifestation of the problem was that input string disappeared when
returning from sleep mode. When editing a TextView with an IME in
landscape orientation, the text would disappear when returning from
sleep mode. The InputMethodManager would be deactivated when the screen
was put into sleep mode as well as the input connection. However when
returning from sleep mode the InputMethodManager was activated, but the
input connection would not be activated again.

The solution is to check focus of the InputMethodManager
which will create a new active input connection to use.

The change is however not specific to this one problem but fundamentally
addresses the issue of lack of startInput on return from sleep mode.

Change-Id: I95d05110bc1cf310fad23ea1bcbc5890f642d1fc